+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1
+
+- Add simple docstring
+- Fix prefix of compressed public key hex
+
 ## 0.1.0
 
 - Pass a basic test